% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{change_trading_env}
\alias{change_trading_env}
\title{change trading environment}
\usage{
change_trading_env()
}
\value{
"live" or "paper" string value
}
\description{
if you wrote live trading and paper trading evironment variables both in
your .Renviron file, you can change trading environment.
}
\examples{
\dontrun{change_trading_environment()}

}
