% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{download_kosdaq_master}
\alias{download_kosdaq_master}
\title{Download kosdaq master file}
\usage{
download_kosdaq_master(base_dir)
}
\arguments{
\item{base_dir}{destination folder, if missing, current working directory.}
}
\value{
no return.
}
\description{
Download kosdaq_code.mst file.
}
\examples{
# download kosdaq_code.mst
\dontrun{download_kosdaq_master()}

}
