% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{download_kospi_master}
\alias{download_kospi_master}
\title{Download kospi master file}
\usage{
download_kospi_master(base_dir)
}
\arguments{
\item{base_dir}{destination folder, if missing, current working directory.}
}
\value{
no return.
}
\description{
Download kospi_code.mst file.
}
\examples{
# download kospi_code.mst
\dontrun{download_kospi_master()}

}
