% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{get_balance}
\alias{get_balance}
\title{get account balance}
\usage{
get_balance(prdt_code, rt_cash_flag = FALSE)
}
\arguments{
\item{prdt_code}{A string specifying account product code}

\item{rt_cash_flag}{A boolean specifying total deposit or not}
}
\value{
balance data.frame
}
\description{
Get your account balance.
}
\examples{
## get account balance
\dontrun{
# detail
get_balance(rt_cash_flag = FALSE)
# total deposit
get_balance(rt_cash_flag = TRUE)}

}
