% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{get_buyable_cash}
\alias{get_buyable_cash}
\title{get buyable cash}
\usage{
get_buyable_cash(stock_code, qry_price = 0, prdt_code)
}
\arguments{
\item{stock_code}{A string specifying stock code}

\item{qry_price}{A numeric or string specifying price}

\item{prdt_code}{A string specifying account product code}
}
\value{
A numeric specifying buyable cash
}
\description{
Get buyable amount of cash of the account
}
\examples{
# get buyable cash
\dontrun{get_buyable_cash("005930")}

}
