% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_current_index}
\alias{get_current_index}
\title{get current index}
\usage{
get_current_index(index_code = c("0001", "1001", "2001"))
}
\arguments{
\item{index_code}{A string specifying an index code (KOSPI: "0001", KOSDAQ: "1001", KOSPI200: "2001")}
}
\value{
a data frame of current index
}
\description{
Get current index.
}
\examples{
# get current index
\dontrun{
get_current_index("0001")
}

}
