% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_price.R
\name{get_current_price}
\alias{get_current_price}
\title{get current price}
\usage{
get_current_price(stock_no)
}
\arguments{
\item{stock_no}{A string specifying stock number (stock code)}
}
\value{
current stock price
}
\description{
Get current stock price using KIS api.
}
\examples{
## get current price
\dontrun{get_current_price()}

}
