% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{get_hash}
\alias{get_hash}
\title{get hash key}
\usage{
get_hash(params)
}
\arguments{
\item{params}{A string specifying params}
}
\value{
A string hash value
}
\description{
Get hash key using KIS app key and KIS app secret.
}
\examples{
# get hash key
\dontrun{get_hash(params)}

}
