% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{get_kosdaq_master_dataframe}
\alias{get_kosdaq_master_dataframe}
\title{get kosdaq master data frame}
\usage{
get_kosdaq_master_dataframe(base_dir)
}
\arguments{
\item{base_dir}{destination folder, if missing, current working directory.}
}
\value{
kosdaq master data frame
}
\description{
Get data from kosdaq_code.mst file.
}
\examples{
# get kosdaq master data frame
\dontrun{kosdaq_master_download()
get_kosdaq_master_dataframe()
}

}
