% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{get_order_history}
\alias{get_order_history}
\title{get order history}
\usage{
get_order_history(sdt, edt, prdt_code, zip_flag = TRUE)
}
\arguments{
\item{sdt}{A string specifying start date "YYYYMMDD"}

\item{edt}{A string specifying end date "YYYYMMDD"}

\item{prdt_code}{A string specifying account product code}

\item{zip_flag}{A boolean specifying choosing important columns}
}
\value{
Order history data frame
}
\description{
Get order history from start date to end date.
}
\examples{
# get order history
\dontrun{
get_order_history("20")
}

}
