% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{get_orders}
\alias{get_orders}
\title{get orders}
\usage{
get_orders()
}
\value{
data.frame of orders
}
\description{
Return a list of orders that can be revised or canceled.
}
\examples{
# get a list of orders
\dontrun{
get_orders()
}

}
