% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_stock_investor}
\alias{get_stock_investor}
\title{get stock investor}
\usage{
get_stock_investor(stock_code)
}
\arguments{
\item{stock_code}{A string specifying stock code}
}
\value{
stock investor data frame
}
\description{
Get stock investor data using KIS API.
}
\examples{
# get stock investor
\dontrun{
get_stock_investor("005930")
}

}
