% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_stock_quotes}
\alias{get_stock_quotes}
\title{get stock quotes}
\usage{
get_stock_quotes(stock_code)
}
\arguments{
\item{stock_code}{A string specifying stock code}
}
\value{
current stock quotes data frame
}
\description{
Get current stock qutoes.
}
\examples{
# get stock quotes
\dontrun{
get_stock_quotes("005930")
}

}
