% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-kospi.R
\docType{data}
\name{kospi_master_columns}
\alias{kospi_master_columns}
\title{Kospi master columns}
\format{
data.frame
}
\usage{
kospi_master_columns
}
\description{
This dataset contains both English and Korean column names
}
\examples{
kospi_master_columns
}
\keyword{datasets}
