% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{print_acnt_prdt_cd}
\alias{print_acnt_prdt_cd}
\title{print account product code}
\usage{
print_acnt_prdt_cd()
}
\value{
account product code
}
\description{
Print your account product code in the current trading environemnt.
it changes depending on the trading environment.
}
\examples{
\dontrun{print_acnt_prdt_cd()}

}
