% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{print_app_key}
\alias{print_app_key}
\title{print app key}
\usage{
print_app_key()
}
\value{
app key
}
\description{
Print your app key in the current trading environemnt.
it changes depending on the trading environment.
}
\examples{
\dontrun{print_app_key()}

}
