% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{print_app_secret}
\alias{print_app_secret}
\title{print app secret}
\usage{
print_app_secret()
}
\value{
app secret
}
\description{
Print your app secret in the current trading environemnt.
it changes depending on the trading environment.
}
\examples{
\dontrun{print_app_secret()}

}
