% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{print_base_url}
\alias{print_base_url}
\title{print base url}
\usage{
print_base_url()
}
\value{
base url of current trading environment
}
\description{
Print your base url in the current trading environemnt.
it changes depending on the trading environment.
}
\examples{
\dontrun{print_base_url()}

}
