% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{print_cano}
\alias{print_cano}
\title{print account number}
\usage{
print_cano()
}
\value{
account number
}
\description{
Print your account number in the current trading environemnt.
it changes depending on the trading environment.
}
\examples{
\dontrun{print_cano()}

}
