% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{print_trading_env}
\alias{print_trading_env}
\title{print trading environment}
\usage{
print_trading_env()
}
\value{
"live" or "paper" string value
}
\description{
Print the current trading environment.
}
\examples{
\dontrun{print_trading_env()}

}
