% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{set_acnt_prdt_cd}
\alias{set_acnt_prdt_cd}
\alias{get_acnt_prdt_cd}
\alias{print_acnt_prdt_cd}
\title{Set KIS account product code}
\usage{
set_acnt_prdt_cd(acnt_prdt_cd)

get_acnt_prdt_cd()

print_acnt_prdt_cd()
}
\arguments{
\item{acnt_prdt_cd}{A string specifying KIS account product code}
}
\value{
No return value, called to set KIS account product code
}
\description{
Save KIS account product code for the current session. To set it permanently,
please add the following line to your .Renvrion file:

KIS_ACNT_PRDT_CD = "YOUR ACCOUNT PRODUCT CODE" (last 2 digits)
}
\examples{

## Set API secret for the current session
\donttest{set_acnt_prdt_cd("your_acnt_prdt_cd")}

## Check API secret
print_acnt_prdt_cd()

}
