% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_app_key}
\alias{set_app_key}
\alias{get_app_key}
\alias{print_app_key}
\title{Set KIS app key}
\usage{
set_app_key(app_key)

get_app_key()

print_app_key()
}
\arguments{
\item{app_key}{A string specifying KIS app key}
}
\value{
No return value, called to set KIS app key
}
\description{
Save KIS app key for the current session. To set it permanently, please add
the following line to your .Renvrion file:

KIS_APP_KEY = "YOUR APP KEY"
}
\examples{

## Set app key for the current session
\donttest{set_app_key("your_app_key")}

## Check app key
print_app_key()

}
