% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_app_secret}
\alias{set_app_secret}
\alias{get_app_secret}
\alias{print_app_secret}
\title{Set KIS app secret}
\usage{
set_app_secret(app_secret)

get_app_secret()

print_app_secret()
}
\arguments{
\item{app_secret}{A string specifying KIS app secret}
}
\value{
No return value, called to set KIS app secret
}
\description{
Save KIS app secret for the current session. To set it permanently, please add
the following line to your .Renvrion file:

KIS_APP_SECRET = "YOUR APP SECRET"
}
\examples{

## Set app secret for the current session
\donttest{set_app_secret("your_app_secret")}

## Check app secret
print_app_secret()

}
