% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_auth}
\alias{set_auth}
\alias{get_auth}
\alias{print_auth}
\title{set access token}
\usage{
set_auth()

get_auth()

print_auth()
}
\description{
Set access token using kis app key and kis app secret.
}
\examples{
# set authentification
\dontrun{set_auth()}

}
