% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{set_cano}
\alias{set_cano}
\alias{get_cano}
\alias{print_cano}
\title{Set KIS common account number}
\usage{
set_cano(cano)

get_cano()

print_cano()
}
\arguments{
\item{cano}{A string specifying KIS common account number}
}
\value{
No return value, called to set KIS common account number
}
\description{
Save KIS common account number for the current session. To set it permanently,
please add the following line to your .Renvrion file:

KIS_CANO = "YOUR COMMON ACCOUNT NUMBER" (first 8 digits)
}
\examples{

## Set API Key for the current session
\donttest{set_cano("your_common_account_number")}

## Check common account number
print_cano()

}
