% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_kis_api_key}
\alias{set_kis_api_key}
\alias{get_kis_api_key}
\alias{print_kis_api_key}
\title{Set KIS api key}
\usage{
set_kis_api_key(api_key)

get_kis_api_key()

print_kis_api_key()
}
\arguments{
\item{api_key}{A string specifying KIS api key}
}
\value{
No return value, called to set kis api key
}
\description{
Save KIS api key for the current session. To set it permanently, please add
the following line to your .Renvrion file:

KIS_API_KEY = "YOUR API KEY"
}
\examples{

## Set API Key for the current session
\donttest{set_kis_api_key("your_api_key")}

## Check API key
print_kis_api_key()

}
