% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_kis_api_secret}
\alias{set_kis_api_secret}
\alias{get_kis_api_secret}
\alias{print_kis_api_secret}
\title{Set KIS api secret}
\usage{
set_kis_api_secret(api_secret)

get_kis_api_secret()

print_kis_api_secret()
}
\arguments{
\item{api_secret}{A string specifying KIS api secret}
}
\value{
No return value, called to set kis api secret
}
\description{
Save KIS api secret for the current session. To set it permanently, please add
the following line to your .Renvrion file:

KIS_API_SECRET = "YOUR API SECRET"
}
\examples{

## Set API secret for the current session
\donttest{set_kis_api_secret("your_api_secret")}

## Check API secret
print_kis_api_secret()

}
