% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{get_buyable_cash}
\alias{get_buyable_cash}
\title{get buyable cash}
\usage{
get_buyable_cash(prdt_code)
}
\arguments{
\item{prdt_code}{A string specifying account product code}
}
\value{
A numeric specifying buyable cash
}
\description{
Get buyable amount of cash of the account
}
\examples{
# get buyable cash
\dontrun{
get_buyable_cash()
}

}
