% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_current_price}
\alias{get_current_price}
\title{get current price}
\usage{
get_current_price(stock_code)
}
\arguments{
\item{stock_code}{A string specifying stock number (stock code)}
}
\value{
current stock price
}
\description{
Get current stock price.
}
\examples{
# get current price
\dontrun{
get_current_price("005930")
}

}
