% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{get_kospi_master_dataframe}
\alias{get_kospi_master_dataframe}
\title{Get kospi master data frame}
\usage{
get_kospi_master_dataframe(base_dir)
}
\arguments{
\item{base_dir}{destination folder, if missing, current working directory.}
}
\value{
kospi master data frame
}
\description{
Get data from kospi_code.mst file.
}
\examples{
# get kospi master data frame
\dontrun{kospi_master_download()
get_kospi_master_dataframe()
}

}
