% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kisopenapi-package.R
\docType{package}
\name{kisopenapi-package}
\alias{kisopenapi}
\alias{kisopenapi-package}
\title{kisopenapi: Korea Investment & Securities Open Trading API}
\description{
To use this package, you will first need to open a Korea investment &
securities account, and get your app key and app secret from the website
\url{https://apiportal.koreainvestment.com/}. Once you have your app key and
app secret, you can save those as environment variables for the current
session using the \code{\link{set_trading_env}} functions.
Alternatively, you can set those permanently by adding the following line to
your .Renviron file:
}
\section{For real trading}{
KIS_CANO="YOUR ACCOUNT NUMBER" (first 8 digits of your account)
KIS_ACNT_PRDT_CD="01" (last 2 digits of your account)
KIS_APP_KEY="YOUR APP KEY"
KIS_APP_SECRET="YOUR APP SECRET"
}

\section{For paper trading}{
KIS_PAPER_CANO="YOUR PAPER ACCOUNT NUMBER" (first 8 digits of your paper account)
KIS_PAPER_ACNT_PRDT_CD="01" (last 2 digits of your paper account)
KIS_PAPER_APP_KEY="YOUR PAPER APP KEY"
KIS_PAPER_APP_SECRET="YOUR PAPER APP SECRET"
Any functions that require your environemt variables try to retrieve those via
\code{Sys.getenv("KIS_CANO")}, \code{Sys.getenv("KIS_ACNT_PRDT_CD")},
\code{Sys.getenv("KIS_APP_KEY")}, \code{Sys.getenv("KIS_APP_SECRET")}
(unless account number, account product code, app key and app secret are
explicitly specified as function arguments).
}

\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
\keyword{internal}
