% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-columns.R
\docType{data}
\name{kospi_master_columns}
\alias{kospi_master_columns}
\alias{kosdaq_master_columns}
\title{kospi and kosdaq master columns}
\format{
data.frame
}
\usage{
kospi_master_columns

kosdaq_master_columns
}
\description{
These datasets contain both English and Korean column names.
}
\examples{
kospi_master_columns
kosdaq_master_columns
}
\keyword{datasets}
