% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_auth}
\alias{set_auth}
\alias{get_auth}
\alias{print_auth}
\title{set access token}
\usage{
set_auth()

get_auth()

print_auth()
}
\value{
An access token (access token is stored in environment variable
at the same time)
}
\description{
Set access token (hash key) using kis app key and kis app secret.
}
\examples{
# set authentification
\dontrun{
set_auth()
}

}
